requirements.txt
pyautogen>=0.10
autogen-ext[openai]
python-dotenv
openai
fastapi
uvicorn[standard]
python-multipart
PyMuPDF
